requirements.txt
about-time==3.1.1
agate==1.7.0
aiosqlite==0.18.0
alembic==1.10.2
alive-progress==2.3.1
altair==5.4.1
amqp==5.1.1
anyio==3.6.2
appnope==0.1.3
apprise==1.3.0
asgi-lifespan==2.0.0
asgiref==3.8.1
astroid==3.3.4
asttokens==2.2.1
async-timeout==4.0.3
asyncpg==0.27.0
attrs==22.2.0
azure-core==1.30.2
azure-storage-blob==12.20.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
betterproto==1.2.5
billiard==3.6.4.0
black==23.3.0
blinker==1.8.2
boto3==1.26.90
botocore==1.29.90
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
cfgv==3.4.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
cloudwatch==1.0.5
codecov==2.1.13
colorama==0.4.6
comm==0.1.4
contourpy==1.2.1
coolname==2.2.0
coverage==7.2.5
croniter==1.3.8
cryptography==39.0.2
cycler==0.12.1
daff==1.3.46
dateparser==1.1.7
db-dtypes==1.2.0
dbt-adapters==1.7.0
dbt-bigquery==1.8.2
dbt-common==1.10.0
dbt-core==1.8.7
dbt-extractor==0.5.1
dbt-postgres==1.8.2
dbt-semantic-interfaces==0.5.1
debugpy==1.8.0
decorator==5.1.1
deepdiff==7.0.1
dill==0.3.6
distlib==0.3.8
Django==4.2
django-cors-headers==3.14.0
django-extensions==3.2.3
django-flags==5.0.13
django-ninja==0.21.0
django-prometheus==2.3.1
djangorestframework==3.14.0
elementary-data==0.15.1
exceptiongroup==1.1.1
executing==1.2.0
Faker==17.6.0
filelock==3.16.1
fonttools==4.53.1
fsspec==2023.1.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.36.0
google-cloud-bigquery==3.23.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.9.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grapheme==0.6.0
graphql-core==3.2.3
greenlet==2.0.2
griffe==0.25.5
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
grpclib==0.4.3
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hologram==0.0.16
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
identify==2.6.1
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.14.0
isodate==0.6.1
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
kombu==5.2.4
kubernetes==26.1.0
lazy-object-proxy==1.9.0
leather==0.3.4
Logbook==1.5.3
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mashumaro==3.14
matplotlib==3.9.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
monotonic==1.6
more-itertools==8.14.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
narwhals==1.13.2
nest-asyncio==1.5.7
networkx==2.8.8
nodeenv==1.9.1
numpy==1.25.2
oauthlib==3.2.2
ordered-set==4.1.0
orjson==3.8.8
packaging==23.0
pandas==2.2.2
paramiko==3.4.0
parsedatetime==2.4
parso==0.8.3
pathspec==0.10.3
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
pluggy==1.0.0
posthog==2.5.0
pre-commit==3.8.0
prometheus-client==0.17.1
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.6
pydeck==0.9.1
pyfiglet==0.8.post1
Pygments==2.14.0
pyhumps==3.8.0
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pymsteams==0.2.2
PyNaCl==1.5.0
pyparsing==3.1.3
pyrsistent==0.19.3
pytest==7.2.2
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-parametrization==2022.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.1.0
ratelimit==2.2.1
readchar==4.0.5
redis==5.0.4
referencing==0.35.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.2
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.6.0
sendgrid==6.10.0
sentry-sdk==2.10.0
sgqlc==16.1
six==1.16.0
slack_sdk==3.29.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sql_metadata==2.11.0
SQLAlchemy==1.4.47
sqlalchemy-bigquery==1.11.0
sqlparse==0.5.1
sshtunnel==0.4.0
stack-data==0.6.2
starkbank-ecdsa==2.2.0
streamlit==1.40.0
stringcase==1.2.0
tabulate==0.9.0
tenacity==9.0.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
tornado==6.3.2
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.12.2
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.21.1
vine==5.0.0
virtualenv==20.23.0
wcwidth==0.2.6
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
